+++ packages/core/botpress/src/dialog/processors.js
@@ -7,9 +7,9 @@ module.exports = {
     send: ({ message, originalEvent, state, flowContext }) => {
       let rendered = message.value
 
-      const additionalData = { state: state }
+      let additionalData = { state }
 
-      if (/{{/i.test(rendered)) {
+      if (rendered && rendered.includes('{{')) {
         rendered = Mustache.render(rendered, {
           ...state,
           event: _.pick(originalEvent, ['raw', 'text', 'type', 'platform', 'user']),
@@ -20,11 +20,12 @@ module.exports = {
         })
       }
 
-      if (/^{.+}$/.test(rendered)) {
-        // Check if it's JSON
-        Object.assign(additionalData, JSON.parse(rendered))
-      } else if (!_.isEmpty(rendered)) {
-        Object.assign(additionalData, { text: rendered })
+      // Check if it's JSON
+      // This form is supposed to be more performant compared to regex
+      if (rendered && rendered[0] === '{' && rendered[rendered.length - 1] === '}') {
+        additionalData = { ...additionalData, ...JSON.parse(rendered) }
+      } else if (rendered) {
+        additionalData = { ...additionalData, text: rendered }
       }
 
       return originalEvent.reply(message.type, additionalData)