+++ packages/core/botpress/src/web/lite.jsx
@@ -12,14 +12,15 @@ import store from './store'
 import { fetchModules } from './actions'
 import InjectedModuleView from '~/components/PluginInjectionSite/module'
 import { moduleViewNames } from '~/util/Modules'
-import { getToken } from '~/util/Auth'
+import { getToken, getUniqueVisitorId } from '~/util/Auth'
 
 const token = getToken()
 if (token) {
   axios.defaults.headers.common['Authorization'] = `Bearer ${token.token}`
 }
 
-const { m, v } = queryString.parse(location.search)
+const { m, v, ref } = queryString.parse(location.search)
+
 const alternateModuleNames = {
   'platform-webchat': 'channel-web'
 }
@@ -28,6 +29,17 @@ const moduleName = alternateModuleNames[m] || m
 class LiteView extends React.Component {
   componentDidMount() {
     this.props.fetchModules()
+    this.sendQueries()
+  }
+
+  sendQueries() {
+    if (!ref) {
+      return
+    }
+
+    const userId = window.__BP_VISITOR_ID || getUniqueVisitorId()
+
+    axios.get(`/api/botpress-platform-webchat/${userId}/reference?ref=${ref}`)
   }
 
   render() {