+++ packages/functionals/botpress-qna/CHANGELOG.md
@@ -3,6 +3,125 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="10.25.0"></a>
+# [10.25.0](https://github.com/botpress/modules/compare/v10.24.4...v10.25.0) (2018-07-23)
+
+
+
+
+**Note:** Version bump only for package @botpress/qna
+
+<a name="10.24.4"></a>
+## [10.24.4](https://github.com/botpress/modules/compare/v10.24.3...v10.24.4) (2018-07-20)
+
+
+
+
+**Note:** Version bump only for package @botpress/qna
+
+<a name="10.24.3"></a>
+## [10.24.3](https://github.com/botpress/modules/compare/v10.24.1...v10.24.3) (2018-07-20)
+
+
+
+
+**Note:** Version bump only for package @botpress/qna
+
+<a name="10.24.2"></a>
+## [10.24.2](https://github.com/botpress/modules/compare/v10.24.1...v10.24.2) (2018-07-20)
+
+
+
+
+**Note:** Version bump only for package @botpress/qna
+
+<a name="10.24.1"></a>
+## [10.24.1](https://github.com/botpress/modules/compare/v10.24.0...v10.24.1) (2018-07-20)
+
+
+
+
+**Note:** Version bump only for package @botpress/qna
+
+<a name="10.24.0"></a>
+# [10.24.0](https://github.com/botpress/modules/compare/v10.23.0...v10.24.0) (2018-07-20)
+
+
+### Features
+
+* **qna:** bp.qna.getQuestion function ([fbbcb23](https://github.com/botpress/modules/commit/fbbcb23))
+* **qna:** qna interception can be customized via hook ([bace4c0](https://github.com/botpress/modules/commit/bace4c0))
+
+
+
+
+<a name="10.23.0"></a>
+# [10.23.0](https://github.com/botpress/modules/compare/v10.22.4...v10.23.0) (2018-07-13)
+
+
+### Bug Fixes
+
+* **qna:** on flow deletion or rename, QNA still works ([efd75d6](https://github.com/botpress/modules/commit/efd75d6))
+
+
+### Features
+
+* **qna:** added "text_redirect" action type ([d8e6825](https://github.com/botpress/modules/commit/d8e6825))
+
+
+
+
+<a name="10.22.4"></a>
+## [10.22.4](https://github.com/botpress/modules/compare/v10.22.3...v10.22.4) (2018-07-11)
+
+
+
+
+**Note:** Version bump only for package @botpress/qna
+
+<a name="10.22.3"></a>
+## [10.22.3](https://github.com/botpress/modules/compare/v10.22.0...v10.22.3) (2018-07-05)
+
+
+
+
+**Note:** Version bump only for package @botpress/qna
+
+<a name="10.22.2"></a>
+## [10.22.2](https://github.com/botpress/modules/compare/v10.22.1...v10.22.2) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package @botpress/qna
+
+<a name="10.22.1"></a>
+## [10.22.1](https://github.com/botpress/modules/compare/v10.22.0...v10.22.1) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package @botpress/qna
+
+<a name="10.22.0"></a>
+# [10.22.0](https://github.com/botpress/modules/compare/v10.21.0...v10.22.0) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package @botpress/qna
+
+<a name="10.21.0"></a>
+# [10.21.0](https://github.com/botpress/modules/compare/v10.17.2...v10.21.0) (2018-07-03)
+
+
+### Features
+
+* **qna:** csv import/export ([394a922](https://github.com/botpress/modules/commit/394a922))
+
+
+
+
 <a name="10.20.1"></a>
 ## [10.20.1](https://github.com/botpress/modules/compare/v10.20.0...v10.20.1) (2018-06-21)